+++ pymodbus/client/base.py
@@ -1,6 +1,7 @@
 """Base for all clients."""
 from __future__ import annotations
 
+import asyncio
 from abc import abstractmethod
 from collections.abc import Awaitable, Callable
 
@@ -57,7 +58,9 @@ class ModbusBaseClient(ModbusClientMixin[Awaitable[ModbusPDU]]):
             self.ctx.comm_params.host,
             self.ctx.comm_params.port,
         )
-        return await self.ctx.connect()
+        rc = await self.ctx.connect()
+        await asyncio.sleep(0.1)
+        return rc
 
     def register(self, custom_response_class: type[ModbusPDU]) -> None:
         """Register a custom response class with the decoder (call **sync**).
@@ -82,6 +85,20 @@ class ModbusBaseClient(ModbusClientMixin[Awaitable[ModbusPDU]]):
         if not self.ctx.transport:
             raise ConnectionException(f"Not connected[{self!s}]")
         return self.ctx.execute(no_response_expected, request)
+
+    def set_max_no_responses(self, max_count: int) -> None:
+        """Override default max no request responses.
+
+        :param max_count: Max aborted requests before disconnecting.
+
+        The parameter retries defines how many times a request is retried
+        before being aborted. Once aborted a counter is incremented, and when
+        this counter is greater than max_count the connection is terminated.
+
+        .. tip::
+            When a request is successful the count is reset.
+        """
+        self.ctx.max_until_disconnect = max_count
 
     async def __aenter__(self):
         """Implement the client with enter block.
@@ -184,6 +201,20 @@ class ModbusBaseSyncClient(ModbusClientMixin[ModbusPDU]):
         if not self.connect():
             raise ConnectionException(f"Failed to connect[{self!s}]")
         return self.transaction.sync_execute(no_response_expected, request)
+
+    def set_max_no_responses(self, max_count: int) -> None:
+        """Override default max no request responses.
+
+        :param max_count: Max aborted requests before disconnecting.
+
+        The parameter retries defines how many times a request is retried
+        before being aborted. Once aborted a counter is incremented, and when
+        this counter is greater than max_count the connection is terminated.
+
+        .. tip::
+            When a request is successful the count is reset.
+        """
+        self.transaction.max_until_disconnect = max_count
 
     # ----------------------------------------------------------------------- #
     # Internal methods