+++ pymodbus/payload.py
@@ -57,7 +57,7 @@ class BinaryPayloadBuilder:
         :param wordorder: The endianness of the word (when wordcount is >= 2)
         :param repack: Repack the provided payload based on BO
         """
-        self.deprecate()
+        # self.deprecate()
         self._payload = payload or []
         self._byteorder = byteorder
         self._wordorder = wordorder
@@ -89,7 +89,7 @@ class BinaryPayloadBuilder:
 
     def encode(self) -> bytes:
         """Get the payload buffer encoded in bytes."""
-        self.deprecate()
+        # self.deprecate()
         return b"".join(self._payload)
 
     def __str__(self) -> str:
@@ -101,7 +101,7 @@ class BinaryPayloadBuilder:
 
     def reset(self) -> None:
         """Reset the payload buffer."""
-        self.deprecate()
+        # self.deprecate()
         self._payload = []
 
     def to_registers(self):
@@ -109,7 +109,7 @@ class BinaryPayloadBuilder:
 
         :returns: The register layout to use as a block
         """
-        self.deprecate()
+        # self.deprecate()
         # fstring = self._byteorder+"H"
         fstring = "!H"
         payload = self.build()
@@ -125,7 +125,7 @@ class BinaryPayloadBuilder:
 
         :returns: The coil layout to use as a block
         """
-        self.deprecate()
+        # self.deprecate()
         payload = self.to_registers()
         coils = [bool(int(bit)) for reg in payload for bit in format(reg, "016b")]
         return coils
@@ -138,7 +138,7 @@ class BinaryPayloadBuilder:
 
         :returns: The payload buffer as a list
         """
-        self.deprecate()
+        # self.deprecate()
         buffer = self.encode()
         length = len(buffer)
         buffer += b"\x00" * (length % 2)
@@ -153,7 +153,7 @@ class BinaryPayloadBuilder:
 
         :param values: The value to add to the buffer
         """
-        self.deprecate()
+        # self.deprecate()
         value = pack_bitstring(values)
         self._payload.append(value)
 
@@ -162,7 +162,7 @@ class BinaryPayloadBuilder:
 
         :param value: The value to add to the buffer
         """
-        self.deprecate()
+        # self.deprecate()
         fstring = self._byteorder + "B"
         self._payload.append(pack(fstring, value))
 
@@ -171,7 +171,7 @@ class BinaryPayloadBuilder:
 
         :param value: The value to add to the buffer
         """
-        self.deprecate()
+        # self.deprecate()
         fstring = self._byteorder + "H"
         self._payload.append(pack(fstring, value))
 
@@ -180,7 +180,7 @@ class BinaryPayloadBuilder:
 
         :param value: The value to add to the buffer
         """
-        self.deprecate()
+        # self.deprecate()
         fstring = "I"
         # fstring = self._byteorder + "I"
         p_string = self._pack_words(fstring, value)
@@ -191,7 +191,7 @@ class BinaryPayloadBuilder:
 
         :param value: The value to add to the buffer
         """
-        self.deprecate()
+        # self.deprecate()
         fstring = "Q"
         p_string = self._pack_words(fstring, value)
         self._payload.append(p_string)
@@ -201,7 +201,7 @@ class BinaryPayloadBuilder:
 
         :param value: The value to add to the buffer
         """
-        self.deprecate()
+        # self.deprecate()
         fstring = self._byteorder + "b"
         self._payload.append(pack(fstring, value))
 
@@ -210,7 +210,7 @@ class BinaryPayloadBuilder:
 
         :param value: The value to add to the buffer
         """
-        self.deprecate()
+        # self.deprecate()
         fstring = self._byteorder + "h"
         self._payload.append(pack(fstring, value))
 
@@ -219,7 +219,7 @@ class BinaryPayloadBuilder:
 
         :param value: The value to add to the buffer
         """
-        self.deprecate()
+        # self.deprecate()
         fstring = "i"
         p_string = self._pack_words(fstring, value)
         self._payload.append(p_string)
@@ -229,7 +229,7 @@ class BinaryPayloadBuilder:
 
         :param value: The value to add to the buffer
         """
-        self.deprecate()
+        # self.deprecate()
         fstring = "q"
         p_string = self._pack_words(fstring, value)
         self._payload.append(p_string)
@@ -239,7 +239,7 @@ class BinaryPayloadBuilder:
 
         :param value: The value to add to the buffer
         """
-        self.deprecate()
+        # self.deprecate()
         fstring = "e"
         p_string = self._pack_words(fstring, value)
         self._payload.append(p_string)
@@ -249,7 +249,7 @@ class BinaryPayloadBuilder:
 
         :param value: The value to add to the buffer
         """
-        self.deprecate()
+        # self.deprecate()
         fstring = "f"
         p_string = self._pack_words(fstring, value)
         self._payload.append(p_string)
@@ -259,7 +259,7 @@ class BinaryPayloadBuilder:
 
         :param value: The value to add to the buffer
         """
-        self.deprecate()
+        # self.deprecate()
         fstring = "d"
         p_string = self._pack_words(fstring, value)
         self._payload.append(p_string)
@@ -269,7 +269,7 @@ class BinaryPayloadBuilder:
 
         :param value: The value to add to the buffer
         """
-        self.deprecate()
+        # self.deprecate()
         fstring = self._byteorder + str(len(value)) + "s"
         self._payload.append(pack(fstring, value.encode()))
 
@@ -302,7 +302,7 @@ class BinaryPayloadDecoder:
         :param byteorder: The endianness of the payload
         :param wordorder: The endianness of the word (when wordcount is >= 2)
         """
-        self.deprecate()
+        # self.deprecate()
         self._payload = payload
         self._pointer = 0x00
         self._byteorder = byteorder
@@ -385,12 +385,12 @@ class BinaryPayloadDecoder:
 
     def reset(self):
         """Reset the decoder pointer back to the start."""
-        self.deprecate()
+        # self.deprecate()
         self._pointer = 0x00
 
     def decode_8bit_uint(self):
         """Decode a 8 bit unsigned int from the buffer."""
-        self.deprecate()
+        # self.deprecate()
         self._pointer += 1
         fstring = self._byteorder + "B"
         handle = self._payload[self._pointer - 1 : self._pointer]
@@ -398,7 +398,7 @@ class BinaryPayloadDecoder:
 
     def decode_bits(self, package_len=1):
         """Decode a byte worth of bits from the buffer."""
-        self.deprecate()
+        # self.deprecate()
         self._pointer += package_len
         # fstring = self._endian + "B"
         handle = self._payload[self._pointer - 1 : self._pointer]
@@ -406,7 +406,7 @@ class BinaryPayloadDecoder:
 
     def decode_16bit_uint(self):
         """Decode a 16 bit unsigned int from the buffer."""
-        self.deprecate()
+        # self.deprecate()
         self._pointer += 2
         fstring = self._byteorder + "H"
         handle = self._payload[self._pointer - 2 : self._pointer]
@@ -414,7 +414,7 @@ class BinaryPayloadDecoder:
 
     def decode_32bit_uint(self):
         """Decode a 32 bit unsigned int from the buffer."""
-        self.deprecate()
+        # self.deprecate()
         self._pointer += 4
         fstring = "I"
         handle = self._payload[self._pointer - 4 : self._pointer]
@@ -423,7 +423,7 @@ class BinaryPayloadDecoder:
 
     def decode_64bit_uint(self):
         """Decode a 64 bit unsigned int from the buffer."""
-        self.deprecate()
+        # self.deprecate()
         self._pointer += 8
         fstring = "Q"
         handle = self._payload[self._pointer - 8 : self._pointer]
@@ -432,7 +432,7 @@ class BinaryPayloadDecoder:
 
     def decode_8bit_int(self):
         """Decode a 8 bit signed int from the buffer."""
-        self.deprecate()
+        # self.deprecate()
         self._pointer += 1
         fstring = self._byteorder + "b"
         handle = self._payload[self._pointer - 1 : self._pointer]
@@ -440,7 +440,7 @@ class BinaryPayloadDecoder:
 
     def decode_16bit_int(self):
         """Decode a 16 bit signed int from the buffer."""
-        self.deprecate()
+        # self.deprecate()
         self._pointer += 2
         fstring = self._byteorder + "h"
         handle = self._payload[self._pointer - 2 : self._pointer]
@@ -448,7 +448,7 @@ class BinaryPayloadDecoder:
 
     def decode_32bit_int(self):
         """Decode a 32 bit signed int from the buffer."""
-        self.deprecate()
+        # self.deprecate()
         self._pointer += 4
         fstring = "i"
         handle = self._payload[self._pointer - 4 : self._pointer]
@@ -457,7 +457,7 @@ class BinaryPayloadDecoder:
 
     def decode_64bit_int(self):
         """Decode a 64 bit signed int from the buffer."""
-        self.deprecate()
+        # self.deprecate()
         self._pointer += 8
         fstring = "q"
         handle = self._payload[self._pointer - 8 : self._pointer]
@@ -466,7 +466,7 @@ class BinaryPayloadDecoder:
 
     def decode_16bit_float(self):
         """Decode a 16 bit float from the buffer."""
-        self.deprecate()
+        # self.deprecate()
         self._pointer += 2
         fstring = "e"
         handle = self._payload[self._pointer - 2 : self._pointer]
@@ -475,7 +475,7 @@ class BinaryPayloadDecoder:
 
     def decode_32bit_float(self):
         """Decode a 32 bit float from the buffer."""
-        self.deprecate()
+        # self.deprecate()
         self._pointer += 4
         fstring = "f"
         handle = self._payload[self._pointer - 4 : self._pointer]
@@ -484,7 +484,7 @@ class BinaryPayloadDecoder:
 
     def decode_64bit_float(self):
         """Decode a 64 bit float(double) from the buffer."""
-        self.deprecate()
+        # self.deprecate()
         self._pointer += 8
         fstring = "d"
         handle = self._payload[self._pointer - 8 : self._pointer]
@@ -496,7 +496,7 @@ class BinaryPayloadDecoder:
 
         :param size: The size of the string to decode
         """
-        self.deprecate()
+        # self.deprecate()
         self._pointer += size
         return self._payload[self._pointer - size : self._pointer]
 
@@ -505,5 +505,5 @@ class BinaryPayloadDecoder:
 
         :param nbytes: The number of bytes to skip
         """
-        self.deprecate()
+        # self.deprecate()
         self._pointer += nbytes